+++ internal/entities/analytics/item.go
@@ -14,3 +14,13 @@ type VoterBucket struct {
 	Votes  string `json:"votes"`
 	Voters uint64 `json:"voters"`
 }
+
+type ExclusiveVoters struct {
+	Count   uint32 `json:"count"`
+	Percent uint32 `json:"percent"`
+}
+
+type ProposalsByMonth struct {
+	PeriodStarted  common.Time `json:"period_started"`
+	ProposalsCount uint64      `json:"proposals_count"`
+}